Warning during compilation on Windows systems
This is based on the warning type: C4702 unreachable code
Another complaint by GCC about
```
warning: comparison of integer expressions of different signedness: ‘int’ and ‘size_t’
```

+++ src/util.cpp
@@ -7255,7 +7255,7 @@ bool recognizeFixedForm(const QCString &contents, FortranFormat format)
   size_t sizCont;
 
   sizCont = contents.length();
-  for (int i=0;i<sizCont;i++)
+  for (size_t i=0;i<sizCont;i++)
   {
     column++;
